[1.2.3] Bugs and formatting
BUGS:
fixed a bug where if you have over 11 columns, it will crash the system. Fixed it by instead truncating the columns. This also allows for notes sections to be added if you really want to.

FORMATTING:
Added "DeviceID":"Modbus1" and "ScheduleID":"1" to the output. Also removed conversion factor, conversion offset, and bitmap if the regtype is descrete or coil.